requirements.txt
streamlit>=1.28.0
openai>=1.0.0
requests>=2.31.0
pyyaml>=6.0
scikit-learn>=1.3.0
plotly>=5.15.0
beautifulsoup4>=4.12.0
networkx>=3.1
matplotlib>=3.7.0
geopandas>=0.13.0
pydeck>=0.8.0
pandas>=2.0.0
pytest>=7.4.0
numpy>=1.24.0
nltk>=3.8.0
spacy>=3.6.0
readability-lxml>=0.8.1
langdetect>=1.0.9
scipy>=1.11.0
wordcloud>=1.9.0